feat: devide ThemeSelectorn into two files: Scheme (dark/light mode) and Theme (color themes) Selectror components
